Numerous backend prov versioning changes #4343
- Commented out cplId for removal or to just lurk until we actually talk to CPL
- Move provEntityName from filemetadata to datafile, as it should not be versioned
- Passing metadata into popup bean for future versioning needs

This work is preparing for versioning of freeform provenance and other changes surrounding the various prov flow.
