+++ scripts/database/upgrades/upgrade_v4.8.5_to_v4.9.sql
@@ -3,5 +3,5 @@ ALTER TABLE externaltool ALTER COLUMN type SET NOT NULL;
 -- Previously, the only explore tool was TwoRavens. We now persist the name of the tool.
 UPDATE guestbookresponse SET downloadtype = 'TwoRavens' WHERE downloadtype = 'Explore';
 ALTER TABLE filemetadata ADD COLUMN prov_freeform text;
-ALTER TABLE filemetadata ADD COLUMN prov_cplid int;
-ALTER TABLE filemetadata ADD COLUMN prov_entityname text;
+-- ALTER TABLE datafile ADD COLUMN prov_cplid int;
+ALTER TABLE datafile ADD COLUMN prov_entityname text;
