+++ src/main/webapp/editFilesFragment.xhtml
@@ -238,7 +238,7 @@
                                         <p:commandLink id="fileProvenanceButton"
                                                    update=":datasetForm:editProvenancePopup"
                                                    oncomplete="PF('editProvenancePopup').show();bind_bsui_components();">
-                                                   <f:actionListener binding="#{provPopupFragmentBean.updatePopupState(fileMetadata.dataFile, EditDatafilesPage.dataset)}" />
+                                            <f:actionListener binding="#{provPopupFragmentBean.updatePopupState(fileMetadata, EditDatafilesPage.dataset)}" />
                                                     #{bundle['file.provenance']} 
                                         </p:commandLink>    
                                     </li>
